code moved to a macro
