+++ frb_example/dart_minimal/lib/main.dart
@@ -1,13 +1,13 @@
 import 'package:frb_example_dart_minimal/src/rust/api/minimal.dart';
 import 'package:frb_example_dart_minimal/src/rust/frb_generated.dart';
 
-// TODO hid these calls
+// TODO hide these calls
 // final hideme2 = FRBLogger.fromLevelFilter();
 
 // final LOGGER = FRBLogger.getLogger('bah');
 // final LOGGER = FRBLogger.init_logger(
 //     custom_log_function: (record) => print(record.message));
-final LOGGER = FRBLogger.init_logger(maxLoglevel: 'SHOUT');
+final LOGGER = FRBLogger.init_logger(maxLoglevel: 'INFO');
 // final LOGGER = FRBLogger.init_logging(maxLoglevel: Level.SHOUT);
 
 // If you are developing a binary program, you may want to put it in `bin/something.dart`
