+++ frb_rust/src/misc/mod.rs
@@ -1,5 +1,7 @@
 pub(crate) mod dart_dynamic;
 pub(crate) mod into_into_dart;
+#[cfg(feature = "log")]
+pub(crate) mod logging;
 pub(crate) mod logs;
 pub(crate) mod manual_impl;
 pub(crate) mod panic_backtrace;
